Ensure ISO has consistent volume ID.
There are a few different implmentations of the `mkisofs` tool. And
there isn't a consistent value for what the default volume ID (i.e.
filesystem label) should be. This change ensures that Azure Linux ISOs
always use the same volume ID regardless of the host they are build on.

This is a bug fix for the Image Customizer Live-CD generator which uses
the ISO's filesystem label in the grub.cfg file.