Image Customizer: Fix ISO gen with SELinux disabled base image.
Currently, the ISO gen disables SELinux in the image because there is
an issue causing the ISO not to boot if SELinux is enabled. However,
the logic to disable SELinux fails if SELinux is already disabled. This
change fixes this problem.

This change also combines the two different implementations for
changing the SELinux mode (in the kernel CLI args). It also makes the
logic more robust to the ordering of the kernel CLI args.