Remove required kernel configs check
This check was added to alert if kernel configs with known required values
have been changed to undesired values or removed. Additionally it would
alert developers to update the json with justification to help with future
checks.

This check is no longer needed now that kernel maintainers are required on
each PR review for configs. Additionally, the check caused confusion
and failed frequently due to the check to update the kernel required
configs json. Therefore, remove.