Add retries for download failure
Add more retries for download failures, retrying up to ~2 minutes with
exponential backoff. Also ensure that the precacher short-circuits the
retries if it receives a 404, since our speculative fetching intentionally
hits these cases.

Also mark precache failures as fatal, since we rely on them for
pinning to an upstream snapshot.