MIC ISO: support for grub-noprefix package use scenario.
When building an Azure Linux image, a user may choose to use the grub-noprefix package. This package installs a grub bootloader with the built-in `prefix` grub env variable set to an empty string. This is typically used for PXE booting where the bootloader and its configuration are placed in the same folder.

The grub-noprefix package installs grubx64-noprefix.efi - but MIC iso generation and the shim (bootx64.efi) both expect a grubx64.efi file. This causes MIC to fail.

This change updates MIC iso generation to:
- look for grubx64-noprefix.efi and understand that it's the same as grubx64.efi (in terms of purpose).
- place grub.cfg file be placed under <iso root>/EFI/BOOT if the noprefix flow is detected. This is basically the folder the bootloader will look for grub.cfg in with this configuration.