openssl: Add dependencies required for tests (#8682)
Change #6932 removed Perl(Test::Harness) and Perl(Math::BigInt) as a build requirement, which breaks tests. Didn't notice this because:

In my test environment, I already had it installed
Our build system ignores check failures
This change adds these back as a build requirement. It also removes an incorrect part of a patch added for SymCrypt provider, which caused tests to fail. This is only a partial fix -- we're working on the right fix but this gets us into a clean state so we can make sure all tests pass from now on.