Image Customizer: Support disk size suffixes.
When specifying the disk and partition sizes, allow the suffixes K, M,
G, and T which mean KiB, MiB, GiB, and TiB respectively.

This change also forces the user to provide a suffix. Previously, the
unsuffixed number meant MiBs. But it should probably be bytes instead.
To help with this transition, unsuffixed numbers have been blocked for
now. But we may support them in the future.