Remove Kernel Required Configs Check (#8661)
This check was added to alert if kernel configs with known required values have been changed to undesired values or removed. Additionally it would alert developers to update the json with justification to help with future checks.

This check is no longer needed now that kernel maintainers are required on each PR review for configs. Additionally, the check caused noise and failed frequently. Therefore, remove.