Image Customizer: Functional tests boilerplate.
A lot of code within the image customizer requires a base image to work
correctly. So, this change adds the option to pass in a core-efi image
file to the tests that can be used as a base image. If the image isn't
provided then the tests that need it will be skipped.

The tests that previously used a fake image have been changed to use
the real base image. And a new test has been added to test partition
customization.