baremetal image: temporarily set selinux to permissive.
With the move to Azure Linux 3.0, the selinux needs to be updated to work
with the new versions of the packages.

To unblock testing and development until the selinux policies are updated,
we are setting selinux to permissive for now.

I've opened a bug to re-enable selinux after fixing the known scenarios:
https://dev.azure.com/mariner-org/ECF/_workitems/edit/7476