New toolchain manifest update from tarball script.
Create a new script that updates the contents of the toolchain manifest
files using the contents of a toolchain archive tarball. This is
intended as a replacement for the existing `update_manifests.sh`
script, with  hopefully better handling the package ordering.
Specifically, it uses the ordering from the existing manifest files
instead of recreating the ordering manually.