Add SymCrypt and SymCrypt-OpenSSL to images. (#8848)
As SymCrypt is meant to be the main crypto provider for openssl in azl3, and the only FIPS-compatible provider. As such, we need both SymCrypt and SymCrypt-OpenSSL in our base images. We can't have openssl have a direct runtime dependency on one of them, because it creates a circular dependency. So instead we have a "recommends", and then explicitly add them to image definitions that have openssl.