Image Customizer: Fix passwords API.
This change refactors the user password API as part of the work to
align the Image Customizer and Trident APIs.

This change also disables the password API in official builds of the
Image Customizer tool. In addition, the docs have been updated to
strongly discourage the use of password API and mentions what
alternatives should be used instead.