Refactor shell execute functions.
This change adds a new `shell.ExecBuilder` type. This type uses the
builder pattern and is used to execute sub-processes. This makes it
easier to add new features to API without experiencing an exponential
explosion in API size, like what can happen with the flat functions.

The `shell.ExecBuilder` type adds the ability to set the working
directory and the sub-process's environment variables. This is needed
for upcoming features in the Image Customizer tool.

All the existing `shell` package flat function have been rewritten to
use the `shell.ExecBuilder` type. Though, some of the functions were
removed in favor of using `shell.ExecBuilder` directly.

In addition, sub-process executions that capture stdout/stderr, can now
also log the stdout/stderr lines. The behavior of `shell.Execute` has
been changed to log stdout to TRACE and stderr to WARN.

Also, the `shell.ExecBuilder` API now makes it possible to set the log
level that the stdout and stderr lines are logged to without
duplicating the lines in TRACE.