Add macro for enabling frame pointers during build and enable frame pointers in glibc (#9092)
Enabling frame pointers during compilation brings benefits to performance profiling and tracing. See https://www.brendangregg.com/blog/2024-03-17/the-return-of-the-frame-pointers.html for more details.

This change in particular adds a macro to add the -fno-omit-frame-pointer flag to CFLAGS when defining %define _include_frame_pointers 1 in the spec file. This way the packager can control enabling whether to build with frame pointers enabled or not. By default, building with frame pointers is disabled.

This change also enables frame pointers by default in our glibc build, which is one of the hot-spots where fast profiling with frame pointers would be beneficial.

Signed-off-by: Chris Co <chrco@microsoft.com>