Image Customizer: Fix cloud-init ISO examples.
There is a behavior in cloud-init where if you use the `seedfrom` arg
on the kernel command-line, then the `network-config` file is ignored.
So, change the example ISO configs that embed cloud-init configs to a
solution that doesn't ignore the `network-config` file.

Also, add a couple of test ISO config files.