Image Customizer: Modify /etc/selinux/config on SELinux disable.
Currently, when the image customizer disables SELinux in an image it
doesn't bother modifying the `/etc/selinux/config` file and instead
only modifies the kernel command-line args. While this is technically
fine, it is confusing for the user. So, this change fixes this so that
`/etc/selinux/config` is kept in sync with the kernel command-line
args.