Image Customizer: Add tests for API package.
This change adds tests for the image customizer API package, so that
the vast majority of lines have test coverage. The only lines remaining
are related to validity checking for kernel command-line args, since
these will very likely be changed/remove in the future (when the API is
changed from a string to a list of strings).

In addition, some of the error messages were cleaned up / made more
consistent.