Disable code optimization for ip frame checksum calculation.
With the current gcc version, the code optimization leads to incorrect checksums
of packets in some cases. This, consequently, leads to those packets getting
dropped by receiving client due to their incorrect checksums. This broke our
PXE boot scenario.

The fix here is a work around where we disable code optimization for the function
in question. Future versions of gcc might fix this issue - at which point, this
fix can be reverted.