Image Customizer: Fix container build.
There is currently an issue where `oras` can't be downloaded when
the container is being built because the container doesn't have the
`ca-certificates` package installed. This problem could be fixed by
adding `ca-certificates` to the list of packages installed. But
instead, this change downloads `oras` binary outside of the
`Dockerfile`. Doing it this way will help reduce the size of the
container.