Image Customizer: Improve resolv.conf behavior.
Instead of just deleting the `resolv.conf` file at the end of
customization, restore it to its original state that was seen in the
base image.

In addition, if the `resolv.conf` didn't exist in the base image and
the `systemd-resolved` service is enabled, then add the symlink that
would normally be created during first-boot. This is particularly
useful for verity images where the filesytem is readonly.

Also, fix the verity example config for Azure Linux 3.0 to have default
network settings that enable a DHCP client on all "physical" NICs. This
is needed since cloud-init doesn't work well in verity images.