Image Customizer: Do not shrink verity hash partition.
The verity hash partition must be given a placeholder filesystem type.
This should probably be `fat32` for simplicty's sake. But currently,
the verity example config uses `ext4`. This causes a problem when the
`--shrink-filesystems` is set because it means the verity hash
partition gets shrunk to almost nothing and therefore isn't big enough
to store the hash tree.

This change fixes this problem by ensuring that the verity hash
partition is never subject to being shrunk regardless of its
placeholder filesystem type. A test is added to verify this.