Image Customizer: Increase loopback detach timeout.
Ubuntu seems to like to call dumpe2fs on ext4 partitions when you mount
a disk (even if you don't mount any of the partitions). Depending the
size of partition, this can take a little bit of time to run, which can
block the loopback from detaching.

The Image Customizer tool uses loopback devices a lot, some of which
are only used for a short period of time. So, it can timeout waiting
for the loopback to detach. So, this change increases the timeout.