Image Customizer: Validate HOME and USER env vars.
The gpg command, and by extension, the tdnf command expects the USER
and HOME environment variables to be valid for the OS they are running
under (including under chroot). Since, the image customizer tool is
typically run under `sudo` and since root is always a valid user, this
generally isn't a problem. But this isn't true if `sudo -E` is used.
And this can result in strange difficult to diganose errors in tdnf.
So, this change verifies that `sudo -E` isn't being used.