Toolkit: Do not give partitions a default label of "primary".
When you call `parted mkpart` for GPT partitions, there is no partition
type parameter but there is a partition label parameter. But currently
the toolkit passes an MBR partition type in the arg that takes the
partition label. This in effect gives all Azure Linux partitions a
default label of "primary", which is a tad silly. This change rectifies
this.