kata-packages-uvm-coco: add e2fsprogs dependency
The new node local CSI driver will require the agent to format a volume
after encrypting it. Hence this adds e2fsprogs to the CoCo UVM so that
the agent can call mkfs.ext4 to format the volume.

A subsequent PR in the microsoft/kata-containers repository will thus
make use of this functionality.

Signed-off-by: Aurélien Bombo <abombo@microsoft.com>