fix: upgrade to 2.11.1.4 and add azurelinux patch
Upgrade WALinuxAgent to the most recent release - 2.11.1.4.
Add patch to include azurelinux support into the agent. The patch should
be removed in the next 2.12 update.
Also because WALinuxAgent will automatically update at runtime, thus
removing the patched-in support for azurelinux, add another patch to
bump the version to 2.11.8.8 so further 2.11 hotfixes are not taken.
Since 2.12 will have azurelinux support, the agent will be able to
automatically upgrade to 2.12 at runtime.

Signed-off-by: Chris Co <chrco@microsoft.com>