fix: use /usr/bin and /usr/lib/systemd paths
In the upstream code's setup.py, the bin path and systemd path are set
through calls into the distro's osutil class. In our case, since we pass
--lnx-distro='mariner', the code will use the MarinerOsUtil derived
class for these paths, which are /usr/bin and /usr/lib/systemd
respectively.

In addition, now that the paths are set correctly again, remove the sed
operation to the waagent.service so the waagent path points to the
correct location again.

Also fix the ephemeral-disk-warning.service to use the proper path as
well.

Signed-off-by: Chris Co <chrco@microsoft.com>