[mic][iso] Add curl to the list of required dependencies for iso creation.
We need to have curl persent on the image to download the rootfs in the PXE
flow. Given that we are aiming to have an iso image whose artifacts can be
used to PXE boot a machine, the PXE scenario is not separate from the iso
scenario and so is its requirements (i.e. curl).