containerized-rpmbuild: switch up strategy for defaults and return command exit codes (#10061)
This change simplifies the containerized-rpmbuild scripts in a few ways, which makes it easier to maintain and simpler to use.

Completely remove azl-3_repo and the associated settings. It ends up being the same as the "real" repo, so it only added complexity.
Switched from using a MACROS environment variable that wrapper functions like rpmspec and rpm had to use in favor of a macros.with-check file that gets copied to the appropriate place and does the same thing without the need to use the environment variable. This allowed me to completely remove the rpmspec wrapper function.
Similar with TDNF_ARGS. Since we no longer need to point to azl-3_repo, all it had was --releasever=3.0. So instead, we install azurelinux-release, which sets that up properly. This allowed me to completely remove the tdnf wrapper function.
Changed the remaining wrapper functions rpm and rpmbuild to return the exit code from the function they wrap, to allow scripts to use the exit code properly.