Add stress-ng to SPECS-EXTENDED (#10102)
Introduce stress-ng into SPECS-EXTENDED. stress-ng is a stress testing tool to exercise the OS & HW and collect metrics. The packaging is initially from Fedora 41.

stress-ng will stress test a computer system in various selectable ways. It was designed to exercise various physical subsystems of a computer as well as the various operating system kernel interfaces.

Signed-off-by: Chris Co <chrco@microsoft.com>