Image Customizer: Account for GPT footer when validating partitions.
For disks partitioned with the GPT partitioning scheme, the last few
sectors of the disk are used for storing the GPT footer (which contains
the backup partition table). The Image Customizer tool does fail if you
try to use this space for a partition, but only because `parted` is
doing the check. This change properly accounts for this footer in the
config validation, so that a nicer error message can be provided to the
user.