fix: remove libssp files to fix avahi hang
avahi-daemon hangs in libssp's fail() routine when built with libssp
support enabled. SSP support is dynamically set when avahi's configure
scans the current build environment for the presence of the standalone
libssp built by gcc.

This standalone implementation of libssp is generally obsoleted in
most modern systems, and instead the libc's implementation of SSP is
used.

So we will remove the libssp files from here if we find they are
present. Avahi's configure will instead use glibc's ssp
implementation which does not hang and is proper.

Signed-off-by: Chris Co <chrco@microsoft.com>