Image Customizer: Always refresh RPM repo metadata. (#10177)
Ensure that when updating/installing packages, the repo metadata is always refreshed to ensure it is up to date. This fixes a bug where running imagecustomizer twice over an image can fail if a local RPM source is updated between the two runs.

In addition, this change ensures that the repo metadata is not refreshed while in the middle of installing/updating the packages.