feat: remove kexec-tools from azure vm definition (#10117)
Azure Linux marketplace image definitions by default should provide all the memory to the user. However since kexec-tools is installed, the crashkernel= boot parameter is added to reserve a certain amount of memory for the kdump kernel. This results in a loss of usable memory for the user, especially if kdump is not armed or being used.

The use of kdump is an opt-in choice by the user, so this change removes pre-loading kexec-tools altogether from the Azure Linux VM images.

Signed-off-by: Chris Co <chrco@microsoft.com>