Fix `gsettings-desktop-schemas` so that the schemas are usable.
The current version of `gsettings-desktop-schemas` does not compile the schemas
after installation, meaning that the schemas aren't available for access by
`gsettings` and other Gnome libraries.  When an application tries to access
schemas that aren't compiled, this can result in error messages like the
following:
```
GLib-GIO-ERROR**: No GSettings schemas are installed on the system
```

This change ensures that compiled schemas are refreshed during post-install and
post-uninstall.

Note that, when not installed by a package manager, the `meson` build script in
`gsettings-desktop-schemas` [will do this in post-install](
https://gitlab.gnome.org/GNOME/gsettings-desktop-schemas/-/blob/master/meson.build?ref_type=heads#L50-52).
Unfortunately, this [doesn't get run in the context of installation via a
package manager because `DESTDIR` is defined](
https://mesonbuild.com/Gnome-module.html#gnomepost_install).