Image Customizer: Be robust to lsblk and fdisk output ordering. (#10199)
It seems that in most cases, the lsblk and fdisk commands output partitions in the correct order. But this isn't actually guaranteed and there have been user reports of unusual errors that are the result of assuming this is the case. This change fixes the code so that it is agnostic to the ordering of the partitions.