fix: change bfq patch to select "none" scheduler as default
On Azure, it is recommended to use an i/o scheduler that passes the scheduling
decisions to the underlying Hyper-V hypervisor. In our case, we should use
the "none" scheduler, which is also ideal for fast random I/O devices like
NVMe. So we update Fedora's bfq patch to change the udev rule to select "none"
instead of Fedora's default Budget Fair Queuing (bfq) and rename the patch
from referencing "bfq" to "none".

Signed-off-by: Chris Co <chrco@microsoft.com>