swap fix-ssl-read-and-write-error-check.patch for a slightly different version from upstream (#10525)
Change #10461 fixed wget2 an upstream issue with openssl error handling using the proposed fix. However, in the end the upstream owner went with a different fix that included a minor refactor.

This change swaps out the original fix for the upstream, to keep us closer to the upstream for future patching.