Add Kdump option to disable reboot after save_core
By default, Kdump reboots the system after the crash dump is saved.
While this works in most cases, it does not work for systems hosted in
RAM where the crash dump is lost once rebooted. To support this
scenario, a new option in the kdump.conf, disable_crash_dump_reboot, is
added and set to 0 by default. Setting this option to 1 will disable the
reboot by having the kdumpctl script exit instead.