Support saving crash dumps on systems hosted in RAM
By default, Kdump reboots the system after the crash dump is saved.
While this works in most cases, it does not work for systems hosted in
RAM where the crash dump is lost once rebooted. To support this
scenario, a new option in the kdump.conf, save_core_to_ram, is
added and set to 0 by default. Setting this option to 1 will disable the
reboot and use the core_collector config option, if specified, to save a
compressed dump. This is necessary because there is no disk to save the
dump to, and there is not enough space in memory to copy the entirety of
the crashed kernel's memory space using the `cp` command. Instead, a
compressed ump must be saved to meet these space limitations.