Update dracut livenet module with an iso downloader script to enable PXE booting for liveos ISOs.
Dracut's livenet module can process root=live:http://artifacts-url at Dracut's parse phase.
However, it does not follow through with downloading the artifacts later - and consequently,
it does not attempt to pivot to the download rootfs image. This blocks the PXE boot flow
for liveos isos.

This change enables this flow by providing a downloader script that is attached to the
livenet module parse phase. That way, once livenet parsing claims the handling of the
kernel root parameter, it can also schedule the downloader to run after the network stack
is up.

To avoid possible conflicts with possible user extensions, the download is only activated
if the new 'rd.live.azldownloader' kernel parameter is present and set to 'enabled' -
otherwise, it is disabled.

Also, the downloader is just a stub that calls into Dracut's livenet root handler which
takes care of downloading.