Image Customizer: Make verity API a list.
Change the verity config from a single item to a list of items. This is
being done so that it is easier to add support for other verity
partitions (e.g. /usr) in the future. However, this change restricts
the verity API to only the root partition (`/`).

In addition, move the verity config from `.os` to `.storage`. This is
being done for alignment with the Trident API. But is also probably a
more morally correct place for verity to be placed.

As a side effect, this change removes support for enabling verity on a
base image that somehow had all the correct partitions to support
verity but didn't actually have verity enabled. None of our base images
are like this. So, it is expected that no user ever made use of this
functionality. This functionality could be re-added in the future. It
was omitted from this change to avoid adding additional complexity.