Toolkit: Add missing `flock` calls.
When making changes to partitions or filesystems, it is recommended to
take a file lock over the disk block device as this informs the host OS
that you are making changes and that it should avoid scanning or
changing the device until you are done. While most of the relevant
operations are covered, there a few places that are missing the lock.
For example, when calling `mkfs` or `resize2fs`.