Image Customizer: Set VHDX block-size to 2 MiB.
Currently, the size of outputted VHDX files can be substantially larger
than VHD or qcow2 files. This behavior occurs because `qemu-img`
dynamically sets the block-size based on the size of the disk. And
larger blocks sizes means the file format is less space efficient,
which results in a larger image file.

This change sets the block-size to 2 MiB for VHDX files. This matches
the `qemu-img` defaults for VHD and QCOW2.

For reference, Windows seems to use a default block-size of 1 MiB for
VHDX files.