Image Customizer: Fix ISO customization.
In change #10789, the `imagecustomizerapi.FileSystem` type had an
internal field added called `PartitionId` which is filled in by the API
validity checks.

ISO to ISO customization supports cracking open the squashfs file and
modifying the Live-OS. However, to do this, a temporary OS image must be
created. This creation logic uses the `imagecustomizerapi.FileSystem`
type but the code wasn't updated to ensure the `PartitionId` field has a
value.