Keep binaries in BOOT and remove fallback efi
In order to preserve consistency with our Azure Linux 3.0 preview and
prior versions of Azure Linux, keep the locations of the binaries in
EFI\BOOT and do not supply a fallback.efi.

This means we should expect the ESP partition to look like:
.
├── EFI
│   ├── BOOT
│   │   ├── bootx64.efi (which is shimx64.efi)
│   │   ├── grubx64.efi
│   │   ├── mmx64.efi

We will re-evaluate this behavior again for the next major OS release.

Signed-off-by: Chris Co <chrco@microsoft.com>