Image Customizer: Use fixed xfs filesystem options.
Currently, we have a fixed set of options that are used when formatting
with the ext4 filesystem type. However, this is currently missing for
xfs. This means that if an Azure Linux 2.0 image is customized in Azure
Linux 3.0, then the OS won't be able to boot since the xfs partition
will have some features enabled that aren't supported on the older
kernel/grub versions.

This change sets the xfs options so that they are consistent regardless
of the OS being used to build/customize the image.