Upgrade mlnx-tools to 24.10.0 version
Building HPC image with infiniband support requires mofed drivers which in turn
requires current version of mlnx-tools. This PR upgrade AZL3.0 mlnx-tools to
the current stable version. It also moves the source tgz to internal blob store.

We do not have to validate bluefield project requirements since it is not active
and hence is not impacted by this change. This was confirmed with Rachel and Juan.

The rpm was installed on a T4 and V100 and verified.