Upgrade mlnx-tools to 24.10.0 version
Building H{C image with infiniband support requires mofed drivers which in turn
requires current version of mlnx-tools. This PR upgrade AZL3.0 mlnx-tools to
the current stable version. It also moves the source tgz to internal blob store.

We do not have to validate bluefield project requirements since it is not active
and hence is no impacted by change. This was confirm with Rachael and Juan.

RPM was installed and verified with T4 and V100 GPUs
 Changes to be committed:
	modified:   SPECS/mlnx-tools/mlnx-tools.signatures.json
	modified:   SPECS/mlnx-tools/mlnx-tools.spec
	modified:   cgmanifest.json