Upgrade runc version to 1.2.2 and libseccomp to 2.5.5 (#11029)
As containerd 2.0 GA recently, we want to provide containerd 2.0 in azure linux 3.0 to enhance our distro. containerd 2.0 requires runc 1.2 as a dependency. The latest runc version 1.2.2 introduces security enhancement and supports new features, such as user namespace, so we want to upgrade runc to the latest version and align with the Fedora offerings. This PR is to upgrade runc version to 1.2.2 and libseccomp to 2.5.5. As a follow-up, containerd 2.0 will be added as a standalone package in a separate PR.