Enable DRM Acceleration and Intel VPU (#11692)
New feature.
This enables the Intel VPU on x86_64. VPU stands for Versatile Processing Unit. This driver is used in the Intel Core Ultra processors to support their AI processing. This driver leans on the DRM Accel framework which requires the Direct Rendering Manager (DRM) subsystem.

Since AI accelerators share many similarities with GPUs, they leverage the existing DRM infrastructure. Therefore as a result of [6e21f45] "Update CONFIG_DRM as loadable module and create sub-package for in-tree amdgpu modules", the accelerator drivers need to be packaged with the subpackage kernel-drivers-gpu. If in the generic kernel package, it will result in errors when insmod (without drm)