glibc: re-enable nscd build and packaging
During initial development of Azure Linux 3.0, the glibc build of nscd was
with no suitable alternative introduced into the distribution for certain
caching, which could typically be fulfilled by sssd.

This change re-enables nscd in the distribution. We will continue to look
at the alternatives during the next major release development cycle.

Signed-off-by: Chris Co <chrco@microsoft.com>