containerd2: provide the capabilities of old-package containerd and obsolete it (#12063)
As containerd2 package is introduced to azurelinux, and containerd2 now is default for AZL 3.0 preview in AKS. There will be conflicts when users install moby-engine in a VM or an AKS node with containerd2 installed as moby-engine requires containerd package. This PR adds "Provides/Obsoletes: containerd" to shift all installs of containerd to containerd2 and resolve the conflicts while installing moby-engine.