openssl speed fixes (#12992)
With the symcrypt provider installed, openssl speed has a couple problems.

First, there are intermittent (but frequent) segfaults in openssl speed ecdh and openssl speed ffdh. This is fixed by upstream pull openssl/openssl#26976. We fix that by taking that patch.

Second, there are several warnings, the root cause of which also makes certain speed test be skipped. This boils down to a feature gap between symcrypt and default provider. For now, we will patch out the specific tests that are warning/bailing, but as those features come online in symcrypt we'll modify/remove that patch.