pam: set config files in /etc as noreplace
When pam package is updated on a system, the package upgrade will
replace the config files found in /etc as well as the /etc/environment
file, even if the user has made edits to the file. As a result, those
edits will be lost after package upgrade.

To solve this, use rpm's %config(noreplace) macro to indicate these
files hold configuration data and will not replace the file if the file
has been altered.

These config macro settings match Fedora's settings.

Signed-off-by: Chris Co <chrco@microsoft.com>