pam: set config files in /etc as noreplace
When pam package is updated on a system, the package upgrade will
replace the /etc/environment file, even if the user has made edits
to the file. As a result, those edits will be lost after package
upgrade.

To solve this, use rpm's %config(noreplace) macro to indicate the
/etc/environment file hold configuration data and will not replace
the file if the file has been altered.

These config macro setting match Fedora's settings for the
/etc/environment file found in the "setup" package.

Signed-off-by: Chris Co <chrco@microsoft.com>