toolkit: enable downloads using managed identities
This change introduces a fallback method for the toolkit's
download functionality. When a download fails and an Azure
client ID was supplied, a fallback attempt is executed if the
download URI is pointing to an Azure Blob Storage.
A new parameter is introduced to supply the Azure client ID.
This new parameter is used for the SRPM packacking functionality,
meaning that the fallback path can be executed for SRPM downloads.
With this, fetching tarballs from authenticated blob storages
(where HTTP GET is expected to fail) is enabled via the fallback
path.

Signed-off-by: Manuel Huber <mahuber@microsoft.com>