osguard: set size to 24 GB to successfully repart
The osguard repartitioning configuration in repart.d ends up
repartitioning the disk to be 23.8 GB in size. The starting image is
only 5.8 GB, so when the image is uploaded to Azure, OS disks created
will start at 6 GB in size, unless directly specified to be greater.

To avoid repartitioning service errors, set the default starting disk
size to an adequate value so repartitioning always is successful.

Signed-off-by: Chris Co <chrco@microsoft.com>