toolkit: enable downloads using managed identities
This change introduces a new method for the toolkit's download
functionality. The new method can be controlled through a new
build variable called 'SRPM_DOWNLOAD_MODE'.
In this change, the new method can be leveraged for the SRPM
packer functionality to source from private storage accounts.
The new method uses Azure CLI credentials and the Azure SDK
for Go to authenticate with these storage accounts.
To use this download mode, end users need to log in to Azure
via Azure CLI. The relevant credentials are then propagated
to the SRPM packer's chroot environment and used by the Azure
SDK for Go.
In the future, this download method can be extended to other
functionalities of the toolkit.

Signed-off-by: Manuel Huber <mahuber@microsoft.com>