[3.0] Passed Azure config to toolkit build env. (#14526)
We're adding support for anonymous Azure blob stores in #14494. This change is to help with that work by making the Azure configuration (including login info) exposed during build-time. Since the toolkit requires to be run as root, the user's Azure config is not visible by default, because it's saved under $HOME/.azure and since we're running as root, the home directory is different from the running user's home directory.