kernel-64k, kernel-hwe: Enable ipmi panic string (#14676)
This change enables the CONFIG_IPMI_PANIC_STRING kernel configuration option. With this option enabled, the kernel will send the panic string to the IPMI system interface during a kernel panic, improving diagnostics and remote debugging capabilities on systems with IPMI support.

This is only changed for kernel-64k and kernel-hwe