Update imagecustomizer to v1.1.0.
Update Azure Linux's Image Customizer to v1.1.0.

In this version, the `go.mod` file was moved to the root of the
git repository. So, there are some minor build fixes to handle this.

Also, fix the version number reported by `imagecustomizer --version`
so that it is only `1.1.0` instead of something like
`1.1.0-dev.20251209.2209+`.

This version has a new dependency on `btrfs-progs` to handle btrfs
filesystems.